Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/browsers/windows/document-access/document_access_parent_access.tentative-expected.txt



FAIL document_access_parent_access assert_equals: Security Error thrown expected "SecurityError" but got "ReferenceError"
